// src/lexer/types/tokenType.ts
export enum TokenType {
  INTEGER = "INTEGER",
  IDENTIFIER = "IDENTIFIER",
  PLUS = "PLUS",
  MINUS = "MINUS",
  MUL = "MUL",
  DIV = "DIV",
  REAL = "REAL",
  MODULE = "MODULE",
  RESERVED_WORD = "RESERVED_WORD",
  ARITHMETIC_OPERATOR = "ARITHMETIC_OPERATOR",
  RELATIONAL_OPERATOR = "RELATIONAL_OPERATOR",
  LOGICAL_OPERATOR = "LOGICAL_OPERATOR",
  ASSIGNMENT_OPERATOR = "ASSIGNMENT_OPERATOR",
  INCREMENT_OPERATOR = "INCREMENT_OPERATOR",
  DECREMENT_OPERATOR = "DECREMENT_OPERATOR",
  PARENTHESIS_OPENING = "PARENTHESIS_OPENING",
  PARENTHESIS_CLOSING = "PARENTHESIS_CLOSING",
  KEY_OPENING = "KEY_OPENING",
  KEY_CLOSING = "KEY_CLOSING",
  BRACKET_OPENING = "BRACKET_OPENING",
  BRACKET_CLOSING = "BRACKET_CLOSING",
  END_OF_STATEMENT = "END_OF_STATEMENT",
  SEPARATOR = "SEPARATOR",
  HEXADECIMAL = "HEXADECIMAL",
  CHARACTER_STRINGS = "CHARACTER_STRINGS",
  BLOCK_COMMENT = "BLOCK_COMMENT",
  INLINE_COMMENT = "INLINE_COMMENT",
  ERROR = "ERROR",
}